Add condition to prevent wiping disk

+++ instance/startup.sh
@@ -1,6 +1,8 @@
 #!/bin/bash 
 sudo mkdir -p /home/minecraft
-sudo mkfs.ext4 -F -E lazy_itable_init=0,lazy_journal_init=0,discard /dev/disk/by-id/google-minecraft-disk
+if [ -z "$(blkid /dev/disk/by-id/google-minecraft-disk)" ]; then
+  sudo mkfs.ext4 -F -E lazy_itable_init=0,lazy_journal_init=0,discard /dev/disk/by-id/google-minecraft-disk
+fi
 sudo mount -o discard,defaults /dev/disk/by-id/google-minecraft-disk /home/minecraft
 cd /home/minecraft
 sudo apt-get update
